Align Diffie-Hellman with current practice
I've updated the last part of DH to be more aligned with current practice.
Static/fixed DH is actively discouraged in practice, so might as well not discuss it.

Simpler solution to fix MitM for DH is to just add signatures.

Corrected the explanation of "ephemeral" and (P)FS, since this does not rely on fixed DH parameters.

Briefly mentioned Elliptic Curve in the context of DH, since ECDHE is the most widely used variant in practice.